jquery invoiceStats
reduced query into one

+++ invoice-server-starter/src/main/java/cz/itnetwork/dto/InvoiceStatisticsDTO.java
@@ -2,10 +2,12 @@ package cz.itnetwork.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class InvoiceStatisticsDTO {
-    private Long currentYearSum;
-    private Long allTimeSum;
+    private BigDecimal currentYearSum;
+    private BigDecimal allTimeSum;
     private Long invoicesCount;
 
 }
